refactor(proxy): move RoutingStrategy into core config
