+++ lmdeploy/serve/proxy/core/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 
-from .config import ProxyConfig
+from .config import ProxyConfig, RoutingStrategy
 from .replica import ReplicaLoad, ReplicaRegistration
 
-__all__ = ['ProxyConfig', 'ReplicaLoad', 'ReplicaRegistration']
+__all__ = ['ProxyConfig', 'ReplicaLoad', 'ReplicaRegistration', 'RoutingStrategy']
